refactor(import): share OneNote types via @triliumnext/commons
The OneNote tree/selection types (OneNoteSection, OneNoteSectionGroup,
OneNoteNotebook, OneNoteFolderRef and the section-selection payload) were
duplicated between the client service and the server Graph client/importer,
risking the two copies drifting apart. Per the repo's shared-types policy,
move them to a single packages/commons/src/lib/onenote.ts so the runImport
payload is type-checked end to end against one definition.

- server graph.ts/importer.ts and the import route now import these from
  commons (the server's SectionSelection is renamed to the shared
  OneNoteSectionSelection); GraphAccount/OneNotePage stay server-internal.
- the client service re-exports them so the dialog and tests keep importing
  from it; OneNoteAccount/OneNoteStatus/OneNoteContainer stay client-side.

Addresses the Greptile review's shared-types finding.
